python/lockerServe/requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
itsdangerous==0.24
pyOSC==0.3.5b-5294
python-osc==1.5
requests==2.7.0
wsgiref==0.1.2
